requirements.txt
flask
gunicorn
google-generativeai
pysqlite3
requests
google-genai